+++ Sem4/OOP/Lab4/services/models/rerq_types/drequest.go
@@ -5,9 +5,9 @@ type DeliveryRequest struct {
 }
 
 type Batch struct {
-	CargoNumber       int     `json:"cargo_number" xml:"cargo_number"`
-	CargoType         string  `json:"cargo_type" xml:"cargo_type"`
-	DeliveryDistance  float64 `json:"delivery_distance" xml:"delivery_distance"`
-	TransportType     string  `json:"transport_type" xml:"transport_type"`
-	TransportInstance string  `json:"transport_instance" xml:"transport_instance"`
+	CargoNumber       int    `json:"cargo_number" xml:"cargo_number"`
+	CargoType         string `json:"cargo_type" xml:"cargo_type"`
+	DeliveryDistance  int    `json:"delivery_distance" xml:"delivery_distance"`
+	TransportType     string `json:"transport_type" xml:"transport_type"`
+	TransportInstance string `json:"transport_instance" xml:"transport_instance"`
 }
